Improve pretty printing of valtrees for references
